feat: Add 4 series support
