+++ epi-display-lg.4Series.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34408.163
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "epi-display-lg.4Series", "src\epi-display-lg.4Series.csproj", "{3076467B-D45F-49C1-B5B9-FB41751E9A47}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3076467B-D45F-49C1-B5B9-FB41751E9A47}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3076467B-D45F-49C1-B5B9-FB41751E9A47}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3076467B-D45F-49C1-B5B9-FB41751E9A47}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3076467B-D45F-49C1-B5B9-FB41751E9A47}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C86729DE-84C2-452B-8912-ED716639AA31}
+	EndGlobalSection
+EndGlobal
