+++ src/Directory.Build.props
@@ -1,0 +1,21 @@
+<Project>
+  <PropertyGroup>
+    <Version>1.0.0-local</Version>
+    <InformationalVersion>$(Version)</InformationalVersion>
+    <Authors>PepperDash Technologies</Authors>
+    <Company>PepperDash Technologies</Company>
+    <Product>PepperDash Crestron LG Display</Product>
+    <Copyright>Copyright ©  2023</Copyright>
+    <RepositoryUrl>https://github.com/PepperDash/epi-lg-display</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <PackageTags>Crestron; 4series</PackageTags>
+    <PackageOutputPath>../output</PackageOutputPath>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+    <PackageLicenseFile>LICENSE.md</PackageLicenseFile>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+  </PropertyGroup>
+  <ItemGroup>
+    <None Include="..\LICENSE.md" Pack="true" PackagePath=""/>
+    <None Include="..\README.md" Pack="true" PackagePath=""/>
+  </ItemGroup>
+</Project>
