+++ src/Properties/AssemblyInfo.cs
@@ -1,0 +1,10 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("Epi-Display-Lg")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("Epi-Display-Lg")]
+[assembly: AssemblyCopyright("Copyright ©  2019")]
+[assembly: AssemblyVersion("0.0.0.*")]
+[assembly: AssemblyInformationalVersion("0.0.0-buildtype-build#")]
+[assembly: Crestron.SimplSharp.Reflection.AssemblyInformationalVersion("0.0.0-buildtype-build#")]
+
