+++ src/epi-display-lg.nuspec
